music/requirements.txt
appdirs==1.4.2
asgiref==1.0.0
attrs==16.3.0
autobahn==0.17.2
Automat==0.5.0
channels==1.0.3
constantly==15.1.0
daphne==1.0.3
Django==1.10.6
django-debug-toolbar==1.6
django-environ==0.4.1
incremental==16.10.1
packaging==16.8
psycopg2==2.5.4
pyparsing==2.1.10
six==1.10.0
Twisted==17.1.0
txaio==2.6.1
zope.interface==4.3.3
